CampaignScreen added and compile issues fixed
A CampaignScreen was added to interact with the user to create a campaign.
Not sure if this is the correct way of doing this but wanted to see with
the rest of you. Also fixed some library issues I had which may have
affected some of the imports in other classes, hope it does not cause too
much problems when merging.