Added reader, writer and rewriter.
ObjectLoader now has a reader method which takes in a path and returns a
list of strings from that text file.
ObjectSaver now has a writer and rewriter depending on if the object has
just been created or has been edited and therefore the file needs to be
rewritten.
These methods are based on the idea that the object files are text files.