Writing to campaign.txt is working
Can now create a new campaign object and save it in the text file. It also
searches for the map the user wants to add and saves the name of the map
in the txt file. If the map does not exist, an exception is thrown and a
message printed stating that map does not exist.