Campaign Creation and test
A new campaign can now not be created with the same name as an existing
campaign. The number of levels is decreased any time a level is removed.
The map names are still missing from the campaign object file and a new
campaign test has been created but needs modifications.