Added a PlayScreen and Play class
The user can now choose a character to play from a list of existing
characters and a campaign to play from an existing list of campaigns.
The PlayScreen also has a getCharacter and a getCampaign methods which
return the chosen character and campaign. The Play class is currently
empty, not sure if we will need it or not.