TODOs added in CampaignScreen
Some TODOs added in order to reduce copied code beteween the screen
classes. There shouldn't be a readLine, readInt and readString methods in
all the screen classes. It would be a good idea to just create a read
class that handles the reading. That class could then close the scanner so
there isn't always a open scanner.