Item validation when editing and creating an Item
The name entered as well as the enhancement type are now validated. The
infinite loop has been fixed. The user can now not create an item with the
same name as an existing item. They cannot edit an item that does not
exist and they cannot add or modify enhancement that does not come with
that item and will be asked to reenter the correct input.