Added a nextLevel method in Campaign
Also added a few methods in Map along with JavaDoc comments. The nextLevel
method gets the next map in a campaign, adds the player to that map, gets
the list of non-player characters in that map and modifies their level,
gets the chest from a map and modifies the enhancement of the items
according to the players level and then returns the map. It also takes the
user back to main menu and tells him he won the game if there are no more
maps in the campaign. The PlayScreen had to be modified because the
nextLevel method takes care of most of the stuff that was there.