Made a lot of changes to Ability
Removed the Strength, Constitution, Dexterity and Modifier classes and
added a value for each in ability instead. Also added a getter and setter
for each. The values of them are being set in the builder. Modified any
instances of them in other classes and added JavaDocs. Removed the level
from the ability class and fixed some methods.